+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ### **(V.3.0.0 Changes) (1.17.0 Minecraft)**
 
+##### Dimension:
+- Min Y has been lowered to -64
+
+##### Biomes:
+- Biomes are larger with Mountainous Blended biome now taller.
+
 ##### Blocks:
 - World Blender Portal Block now has a new overlay when inside the block!